+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Renamed from Attributes.js to DDL.js.
+
 ## 0.1.337 (Oct 4, 2013)
 - First release.  
   And the most overused cheesy word award of this year goes to: automagic.